separate common functions into mixin class
